First pass iroh discovery
